fix: Entity 아이디 타입 통일

+++ src/main/java/org/course/registration/entity/Enroll.java
@@ -17,8 +17,9 @@ import jakarta.persistence.*;
 @Getter
 public class Enroll {
 
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "student_id")
